+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Features:
 - Add support for CMakePresets V7 and V8. [#3549](https://github.com/microsoft/vscode-cmake-tools/issues/3549)
 - Update `api.ts` to add the `onSelectedConfigurationChanged` event. [#3671](https://github.com/microsoft/vscode-cmake-tools/pull/3671) [@OrkunTokdemir](https://github.com/OrkunTokdemir)
 - Improve CMake QuickStart by allowing the user to dynamically create a CMakePresets.json file. [#3649](https://github.com/microsoft/vscode-cmake-tools/issues/3649)
+- Add support to configure default project in workspace setting. [#1078](https://github.com/microsoft/vscode-cmake-tools/issues/1078)
 
 Improvements: